% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctcovidstop-package.R
\docType{package}
\name{ctcovidstop-package}
\alias{ctcovidstop}
\alias{ctcovidstop-package}
\title{ctcovidstop: Clinical trials that stopped during the Covid-19 pandemic}
\usage{
data(c19stoppedtrials)
}
\description{
Provides a database of all ClinicalTrials.gov NCT Numbers corresponding to clinical trials that were "stopped" (had their overall status changed to "Terminated," "Suspended," or "Withdrawn") after 2019-12-01. This dataset indicates the date that a trial was stopped, whether it was started again and on what date, and the contents of the "why stopped?" field on the date the trial was stopped. This dataset also includes columns with manually coded data for whether the "why stopped?" field explicitly indicates that the reason for stopping included the SARS-CoV-2 pandemic.
}
\details{
This data set was last updated on 2022-12-05.
}
\references{
Carlisle, B.G. Clinical Trials Stopped by Covid-19.
https://covid19.bgcarlisle.com
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bgcarlisle/ctcovidstop}
  \item Report bugs at \url{https://github.com/bgcarlisle/ctcovidstop/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Gregory Carlisle \email{murph@bgcarlisle.com} (\href{https://orcid.org/0000-0001-8975-0649}{ORCID})

}
\keyword{internal}
